Reference Microsoft.AspNetCore.App for DependencyInjection.
This makes it a lot easier as then the DependencyInjection package (being the same version as the one it it's framework) can be replaced and get free updates from the one in the framework. So to prevent possible security issues later, always use the one from the framework instead of the package version.

Signed-off-by: AraHaan <seandhunt_7@yahoo.com>